// src/event-form/__test__/data.d.ts
export const spaceTags = [
  {
    id: 'JnG40OGz',
    name: '是',
    creator: 'haozhigang',
    creatorId: 'rRGoYQml',
    createdAt: '2021-01-14T07:08:25Z',
    updater: 'haozhigang',
    updaterId: 'rRGoYQml',
    updatedAt: '2021-01-14T07:08:25Z',
    description: '是  得到的',
    platforms: ['minp'],
    docType: 'page',
    actions: ['page'],
    patternMatched: false,
    attrs: {
      content: null,
      index: null,
      href: null,
      path: 'pages/index/index',
      pg: null,
      query: null,
      xpath: null,
      contentType: null,
      domain: 'wx123456',
      urlScheme: null,
      __typename: 'DocProps',
    },
    definition: {
      domain: 'wx123456',
      path: 'pages/index/index',
      query: '得到的=s',
      xpath: null,
      index: null,
      href: null,
      content: null,
      pg: null,
      contentType: null,
      urlScheme: null,
      __typename: 'DocProps',
    },
    screenshot: { target: '', viewport: '', __typename: 'Screenshot' },
    __typename: 'Element',
  },
];
export const deviceInfoMinp = { name: '小程序无埋点', domain: 'wx123456', os: 'minp' };